move gendoc to gulp/util folder
