+++ gulp/tasks/doc.js
@@ -1,6 +1,6 @@
 var gulp = require('gulp');
 
-var gendoc = require('../../docbuilder/gendoc');
+var gendoc = require('../util/gendoc');
 var config = require('../../build/config');
 
 gulp.task('doc', function () {
